moved component to notifications
